end work module ( insert job, insert job, insert job prices, update men and work, cancel job )

+++ app/src/main/java/com/rojasdev/apprecconproject/data/dataModel/workTotalCollector.kt
@@ -3,6 +3,6 @@ package com.rojasdev.apprecconproject.data.dataModel
 data class workTotalCollector(
     val PK_ID_Recolector: Int,
     val name_recolector: String?,
-    val days_work: Double,
+    val days_work: Int,
     val total: Double
 )
